+++ domain/mathrank-auth-domain/src/main/java/kr/co/mathrank/domain/auth/service/MemberRegisterService.java
@@ -8,11 +8,9 @@ import org.springframework.validation.annotation.Validated;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import kr.co.mathrank.common.snowflake.Snowflake;
-import kr.co.mathrank.domain.auth.dto.MemberInfoCompleteCommand;
 import kr.co.mathrank.domain.auth.dto.MemberRegisterCommand;
 import kr.co.mathrank.domain.auth.entity.Member;
 import kr.co.mathrank.domain.auth.exception.AlreadyExistLoginIdException;
-import kr.co.mathrank.domain.auth.exception.CannotFoundMemberException;
 import kr.co.mathrank.domain.auth.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -39,21 +37,6 @@ public class MemberRegisterService {
 
 		log.info("[MemberRegisterService.register] member saved - memberId: {}", newMember.getId());
 		return newMember.getId();
-	}
-
-	@Transactional
-	public void completeRegister(@NotNull @Valid final MemberInfoCompleteCommand command) {
-		memberRepository.findById(command.memberId())
-			.ifPresentOrElse(member -> {
-				// 사용자 존재 시, 업데이트 마무리
-					member.completeRegister(command.memberType(), command.agreeToPrivacyPolicy(), command.schoolCodes());
-				},
-				() -> {
-				// 사용자를 찾을 수 없을 때, 예외처리
-					log.warn("[MemberRegisterService.completeRegister] cannot find member: {}", command.memberId());
-					throw new CannotFoundMemberException();
-				});
-		log.info("[MemberRegisterService.completeRegister] member register completed - memberId: {}", command.memberId());
 	}
 
 	private boolean isExist(final String loginId) {
